Refresh patches.suse/net-prevent-mss-overflow-in-skb_segment.patch.
Fix the following warning:
net/core/skbuff.c: In function 'skb_segment':
include/linux/kernel.h:795:16: warning: comparison of distinct pointer types lacks a cast [enabled by default]
include/linux/kernel.h:798:2: note: in expansion of macro '__min'
net/core/skbuff.c:3302:18: note: in expansion of macro 'min'

This is how the warning got silenced in upstream stable kernel
v4.19.321.

suse-commit: 68ad1eabfa016d3cd7d9714644e368b13e27b332

+++ net/core/skbuff.c
@@ -3304,7 +3304,7 @@ struct sk_buff *skb_segment(struct sk_buff *head_skb,
 		 * now.
 		 * Cap len to not accidentally hit GSO_BY_FRAGS.
 		 */
-		partial_segs = min(len, GSO_BY_FRAGS - 1) / mss;
+		partial_segs = min(len, (unsigned int)(GSO_BY_FRAGS - 1)) / mss;
 		if (partial_segs > 1)
 			mss *= partial_segs;
 		else
